+++ tests/test_descriptions.py
@@ -21,12 +21,13 @@ import unittest
 from robot_descriptions import (
     allegro_hand_description,
     atlas_description,
+    baxter_description,
     cf2_description,
     double_pendulum_description,
     edo_description,
     finger_edu_description,
-    iiwa_description,
     gen2_description,
+    iiwa_description,
     panda_description,
     pr2_description,
     upkie_description,
@@ -49,6 +50,7 @@ class TestDescriptions(unittest.TestCase):
         descriptions = [
             allegro_hand_description,
             atlas_description,
+            baxter_description,
             cf2_description,
             double_pendulum_description,
             edo_description,