+++ robot_descriptions/_description_names.py
@@ -39,6 +39,7 @@ DESCRIPTION_NAMES: List[str] = [
     "icub_description",
     "iiwa_description",
     "jvrc_description",
+    "minitaur_description",
     "panda_description",
     "pr2_description",
     "reachy_description",