+++ robot_descriptions/_description_names.py
@@ -41,6 +41,7 @@ DESCRIPTION_NAMES: List[str] = [
     "iiwa_description",
     "jvrc_description",
     "laikago_description",
+    "mini_cheetah_description",
     "minitaur_description",
     "panda_description",
     "pr2_description",