+++ robot_descriptions/solo_description.py
@@ -16,17 +16,17 @@
 # limitations under the License.
 
 """
-Laikago description.
+Solo description.
 """
 
 from os import path as _path
 
 from ._cache import clone_to_cache as _clone_to_cache
 
-__working_dir__ = _clone_to_cache("unitree_ros")
+__working_dir__ = _clone_to_cache("example-robot-data")
 
-PATH: str = _path.join(__working_dir__, "robots", "laikago_description")
+PATH: str = _path.join(__working_dir__, "robots", "solo_description")
 
 MESHES_PATH: str = _path.join(PATH, "meshes")
 
-URDF_PATH: str = _path.join(PATH, "urdf", "laikago.urdf")
+URDF_PATH: str = _path.join(PATH, "robots", "solo12.urdf")