+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### New models
 
+- Allegro Hand
 - Crazyflie 2.0
 - Double Pendulum
 - e.DO