+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Allegro Hand
 - Atlas
 - Baxter
+- Bolt
 - Crazyflie 2.0
 - Double Pendulum
 - e.DO