+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - iCub
 - iiwa 14
 - JVRC-1
+- Minitaur
 - Panda
 - PR2
 - Reachy