+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to this project will be documented in this file.
 - Reachy
 - Romeo
 - Simple Humanoid
+- Solo
 - TALOS
 - TIAGo
 - Upkie